+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/persistence/entities/WorkbookDescriptorEntity.kt
@@ -1,0 +1,8 @@
+package org.wycliffeassociates.otter.jvm.workbookapp.persistence.entities
+
+data class WorkbookDescriptorEntity(
+    var id: Int,
+    var sourceFk: Int,
+    var targetFk: Int,
+    var typeFk: Int
+)
